requirements.txt
pyenchant
pytesseract
nltk
